Add matrix_divided and say_my_name modules with input validation and error handling

+++ python-test_driven_development/0-add_integer.py
@@ -20,4 +20,4 @@ def add_integer(a, b=98):
         raise TypeError("a must be an integer")
     if type(b) not in [int, float]:
         raise TypeError("b must be an integer")
-    return int(a) + int(b)
+    return int(float(a) + float(b))
